+++ src/components/ToggleBox/index.tsx
@@ -1,5 +1,5 @@
 import Toggle from '@components/Toggle'
-import { Typo14, Typo15, Typo17 } from '@components/Typo/Typo'
+import { Typo14, Typo15, Typo17 } from '@components/Typo'
 import { SETTINGS_INFO } from '@constants/settingsInfo'
 import { SettingsStoreKey } from '@stores/settingsStore'
 import * as S from './styles'
